fix some overflow and add figsize to python script

+++ src/lib/placement.cpp
@@ -26,7 +26,7 @@ void fillSegment(vector <vector <int>> &PlacementState, int row, int left_pointe
 //if flag = 0, place the bottom die, if flag = 1, place the top die
 void InitializePlacement(Die *currentDie, TopBottomCellArray *ArrayInfo, int flag, bool *PartitionAgain){
     *PartitionAgain = true;
-    int usingCell = 0;
+    unsigned long long int usingCell = 0;
     vector <vector <int>> PlacementState(currentDie->repeatCount,vector <int>(currentDie->rowLength,EMPTY_STATE));
     int sucessful_placement_count = 0;
     int currentCellNumber = 0;
@@ -73,8 +73,8 @@ void InitializePlacement(Die *currentDie, TopBottomCellArray *ArrayInfo, int fla
     if(sucessful_placement_count == currentCellNumber) *PartitionAgain = false;
     printf("successful placement count: %d \n", sucessful_placement_count);
 
-    printf("MaxUtil of bottom die: %d Current Partition Util: %d \n\n", currentDie->MaxUtil, usingCell *100/rowLength / currentDie->repeatCount);
-    if(usingCell * 100/currentDie->rowLength / currentDie->repeatCount > currentDie->MaxUtil) *PartitionAgain = true;
+    printf("MaxUtil of bottom die: %d Current Partition Util: %llu \n\n", currentDie->MaxUtil, usingCell *100/rowLength / currentDie->repeatCount);
+    if(usingCell * 100/currentDie->rowLength / currentDie->repeatCount > (unsigned long long int)currentDie->MaxUtil) *PartitionAgain = true;
 
     //pack data
     if(flag == 0){
